Changed datatypes to double.
The original point, rect classes were moved to the ui lib. It makes sense
to use ints for screen related coordinates and doubles for math. The only
other alternative is to use doubles for both, and we may return to that at
some point.

+++ src/com/cjmalloy/mathlib/shared/linear/Point.java
@@ -5,12 +5,12 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 public class Point implements IsSerializable
 {
 
-    public int x, y;
+    public double x, y;
 
     public Point()
     {}
 
-    public Point(int x, int y)
+    public Point(double x, double y)
     {
         this.x = x;
         this.y = y;
@@ -21,7 +21,7 @@ public class Point implements IsSerializable
         this(p.x, p.y);
     }
 
-    public Point add(int x, int y)
+    public Point add(double x, double y)
     {
         this.x += x;
         this.y += y;
