+++ src/com/cjmalloy/mathlib/shared/linear/Rect.java
@@ -5,15 +5,15 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 public class Rect implements IsSerializable
 {
 
-    public int x;
-    public int y;
-    public int width;
-    public int height;
+    public double x;
+    public double y;
+    public double width;
+    public double height;
 
     public Rect()
     {}
 
-    public Rect(int x, int y, int width, int height)
+    public Rect(double x, double y, double width, double height)
     {
         this.x = x;
         this.y = y;
@@ -28,8 +28,8 @@ public class Rect implements IsSerializable
 
     public Rect boundingRect(Rect r)
     {
-        int ux = Math.min(r.x, x);
-        int uy = Math.min(r.y, y);
+        double ux = Math.min(r.x, x);
+        double uy = Math.min(r.y, y);
         width = Math.max(r.x + r.width, x + width) - ux;
         height = Math.max(r.y + r.height, y + height) - uy;
         x = ux;
@@ -37,7 +37,7 @@ public class Rect implements IsSerializable
         return this;
     }
 
-    public boolean contains(int x, int y)
+    public boolean contains(double x, double y)
     {
         return x > this.x &&x < this.x + this.width &&
                 y > this.y && y < this.y + this.height;
